ColorizedFractal

+++ src/main/kotlin/ru/smak/math/fractals/AlgebraicFractal.kt
@@ -3,5 +3,5 @@ package ru.smak.math.fractals
 import org.kotlinmath.Complex
 
 interface AlgebraicFractal {
-    fun isInSet(c: Complex): Float
+    fun isInSet(c: Complex): Double
 }
